requirements.txt
Package            Version
------------------ -------
imageio            2.33.1
numpy              1.26.4
pillow             10.2.0
pip                22.0.2
py3exiv2           0.12.0
PySide6            6.6.1
PySide6-Addons     6.6.1
PySide6-Essentials 6.6.1
rawpy              0.19.0
setuptools         59.6.0
shiboken6          6.6.1
